Feature/2790 port names performance (#2824)
* Improved performance of Address.IsZero

* Improved base.Address performance

* Removed memory cache from Address
