+++ src/apps/chifra/pkg/types/types_test.go
@@ -1,7 +1,6 @@
 package types
 
 import (
-	"fmt"
 	"testing"
 
 	base "github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/base"
@@ -30,62 +29,5 @@ func TestTypes(t *testing.T) {
 
 	if data["blockNumber"] == nil {
 		t.Fatal("visible field missing when showHidden = true")
-	}
-}
-
-func TestAddress_Hex(t *testing.T) {
-	hex := "0xF1AA581F353005BA3765B81BF52D6B1C488C2101"
-	var addr base.Address
-	addr.SetHex(hex)
-
-	expected := "0xf1aa581f353005ba3765b81bf52d6b1c488c2101"
-	if result := addr.Hex(); result != expected {
-		t.Fatal("wrong Hex() return value:", result)
-	}
-
-	zero := "0x0000000000000000000000000000000000000000"
-	expected = "0x0"
-	addr = base.HexToAddress(zero)
-	if result := addr.Hex(); result != expected {
-		t.Fatal("wrong Hex() return value for 0x0:", result)
-	}
-
-	var zero2 base.Address
-	addrStr := fmt.Sprint(zero2)
-	if addrStr != expected {
-		t.Fatal("wrong Hex() return value for 0x0:", addrStr)
-	}
-}
-
-func TestAddress_Stringer(t *testing.T) {
-	hex := "0xF1AA581F353005BA3765B81BF52D6B1C488C2101"
-	var addr base.Address
-	addr.SetHex(hex)
-
-	expected := "0xf1aa581f353005ba3765b81bf52d6b1c488c2101"
-	if result := fmt.Sprint(addr); result != expected {
-		t.Fatal("wrong Hex() return value:", result)
-	}
-}
-
-func TestAddress_IsZero(t *testing.T) {
-	var zeroValue base.Address
-	if result := zeroValue.IsZero(); result != true {
-		t.Fatal("wrong result for zero value")
-	}
-
-	zeroAddr := base.HexToAddress("0x0")
-	if result := zeroAddr.IsZero(); result != true {
-		t.Fatal("wrong result for zero address")
-	}
-}
-
-func TestHexToAddress(t *testing.T) {
-	hex := "0xF1AA581F353005BA3765B81BF52D6B1C488C2101"
-	addr := base.HexToAddress(hex)
-
-	expected := "0xf1aa581f353005ba3765b81bf52d6b1c488c2101"
-	if result := addr.Hex(); result != expected {
-		t.Fatal("HexToAddress: wrong Hex() return value:", result)
 	}
 }
